+++ test/API/H5_api_file_test.c
@@ -1713,7 +1713,7 @@ test_get_file_obj_count(void H5_ATTR_UNUSED *params)
 
             if (obj_count != 2) {
                 H5_FAILED();
-                printf("    number of open files (%ld) did not match expected number (2)\n", obj_count);
+                printf("    number of open files (%ld) did not match expected number (2)\n", (long)obj_count);
                 PART_ERROR(H5Fget_obj_count_files);
             }
 
@@ -1734,7 +1734,7 @@ test_get_file_obj_count(void H5_ATTR_UNUSED *params)
 
             if (obj_count != 1) {
                 H5_FAILED();
-                printf("    number of open groups (%ld) did not match expected number (1)\n", obj_count);
+                printf("    number of open groups (%ld) did not match expected number (1)\n", (long)obj_count);
                 PART_ERROR(H5Fget_obj_count_grps_single_file);
             }
 
@@ -1755,7 +1755,7 @@ test_get_file_obj_count(void H5_ATTR_UNUSED *params)
 
             if (obj_count != 1) {
                 H5_FAILED();
-                printf("    number of open groups (%ld) did not match expected number (1)\n", obj_count);
+                printf("    number of open groups (%ld) did not match expected number (1)\n", (long)obj_count);
                 PART_ERROR(H5Fget_obj_count_grps);
             }
 
@@ -1777,7 +1777,7 @@ test_get_file_obj_count(void H5_ATTR_UNUSED *params)
             if (obj_count != 1) {
                 H5_FAILED();
                 printf("    number of open named datatypes (%ld) did not match expected number (1)\n",
-                       obj_count);
+                       (long)obj_count);
                 PART_ERROR(H5Fget_obj_count_types);
             }
 
@@ -1798,7 +1798,7 @@ test_get_file_obj_count(void H5_ATTR_UNUSED *params)
 
             if (obj_count != 1) {
                 H5_FAILED();
-                printf("    number of open attributes (%ld) did not match expected number (1)\n", obj_count);
+                printf("    number of open attributes (%ld) did not match expected number (1)\n", (long)obj_count);
                 PART_ERROR(H5Fget_obj_count_attrs);
             }
 
@@ -1819,7 +1819,7 @@ test_get_file_obj_count(void H5_ATTR_UNUSED *params)
 
             if (obj_count != 1) {
                 H5_FAILED();
-                printf("    number of open datasets (%ld) did not match expected number (1)\n", obj_count);
+                printf("    number of open datasets (%ld) did not match expected number (1)\n", (long)obj_count);
                 PART_ERROR(H5Fget_obj_count_dsets);
             }
 
@@ -1841,7 +1841,7 @@ test_get_file_obj_count(void H5_ATTR_UNUSED *params)
             /* One for the file and another for the group */
             if (obj_count != 2) {
                 H5_FAILED();
-                printf("    number of open objects (%ld) did not match expected number (2)\n", obj_count);
+                printf("    number of open objects (%ld) did not match expected number (2)\n", (long)obj_count);
                 PART_ERROR(H5Fget_obj_count_all_single_file);
             }
 
@@ -1862,7 +1862,7 @@ test_get_file_obj_count(void H5_ATTR_UNUSED *params)
 
             if (obj_count != 6) {
                 H5_FAILED();
-                printf("    number of open objects (%ld) did not match expected number (6)\n", obj_count);
+                printf("    number of open objects (%ld) did not match expected number (6)\n", (long)obj_count);
                 PART_ERROR(H5Fget_obj_count_all);
             }
 
@@ -2021,7 +2021,7 @@ test_file_open_overlap(void H5_ATTR_UNUSED *params)
 
     if (obj_count != 3) {
         H5_FAILED();
-        printf("    number of objects opened in file (%ld) did not match expected number (3)\n", obj_count);
+        printf("    number of objects opened in file (%ld) did not match expected number (3)\n", (long)obj_count);
         goto error;
     }
 
@@ -2050,7 +2050,7 @@ test_file_open_overlap(void H5_ATTR_UNUSED *params)
     if (obj_count != 2) {
         H5_FAILED();
         printf("    number of objects opened in the file (%ld) did not match expected number (2)\n",
-               obj_count);
+               (long)obj_count);
         goto error;
     }
